requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==5.0.1
auto-labeling-pipeline==0.1.21
billiard==3.6.4.0
boto3==1.26.142
botocore==1.29.142
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.8
Django==4.2.1
django-allauth==0.50.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-drf-filepond==0.4.1
django-filter==21.1
django-health-check==3.17.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
doccano==1.8.3
drf-yasg==1.21.5
environs==9.5.0
et-xmlfile==1.1.0
filetype==1.2.0
flower==1.2.0
furl==2.1.3
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kombu==5.2.4
lml==0.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging==23.1
pandas==1.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyJWT==2.7.0
python-chess==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==6.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
seqeval==1.2.2
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
stockfish==3.28.0
texttable==1.6.7
threadpoolctl==3.1.0
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
whitenoise==6.4.0
psycopg[binary]
djangorestframework
djangorestframework-simplejwt